place order using products from db by all users

+++ src/main/java/com/practice/onlineShop/repositories/ProductRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.Optional;
 
-public interface ProductRepository extends CrudRepository<Product, Double> {
+public interface ProductRepository extends CrudRepository<Product, Integer> {
 
     Optional<Product>findByCode(String productCode);
 }
